Reorganize src/maccabipediabot/ by sport and function
New structure:
- common/   — shared utilities (pywikibot_boilerplate, prettify_games_pages,
               maccabistats_player_event)
- football/ — gamesbot, upload_last_game, fetch_games, update_league_table,
               sort_players_events, refresh_games
- volleyball/ — all volleyball_* modules + gamesbot_volleyball, upload,
                 update_league_table, reorder, reorganize, create_redirects
- basketball/ — expanded with gamesbot_basketball, update tables,
                 upload scripts, fix_template, create_redirects
- videos/   — merged active/notify_broken_youtube_videos here; removed active/
- maintenance/ — one-off fix/move scripts (move_*, refresh_by_category)

Update all imports and test imports accordingly.
All 21 tests pass.
