+++ src/maccabipediabot/basketball/fix_basketball_players_template.py
@@ -3,7 +3,7 @@ import logging
 import mwparserfromhell
 from pywikibot import pagegenerators
 
-from maccabipediabot.pywikibot_boilerplate import run_boilerplate
+from maccabipediabot.common.pywikibot_boilerplate import run_boilerplate
 
 logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.DEBUG)
