+++ src/maccabipediabot/basketball/refereesbot_basketball.py
@@ -22,7 +22,7 @@ import sys
 
 import requests
 
-from maccabipediabot.pywikibot_boilerplate import run_boilerplate
+from maccabipediabot.common.pywikibot_boilerplate import run_boilerplate
 
 logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)
 logger = logging.getLogger(__name__)
