+++ src/maccabipediabot/basketball/upload_basketball_games_from_files.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import pywikibot as pw
 
-from maccabipediabot.pywikibot_boilerplate import run_boilerplate
+from maccabipediabot.common.pywikibot_boilerplate import run_boilerplate
 
 logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.DEBUG)
