+++ src/maccabipediabot/basketball/upload_basketball_players.py
@@ -2,7 +2,7 @@ import logging
 
 import pywikibot as pw
 
-from maccabipediabot.pywikibot_boilerplate import run_boilerplate
+from maccabipediabot.common.pywikibot_boilerplate import run_boilerplate
 
 logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.DEBUG)
