+++ src/maccabipediabot/common/maccabistats_player_event.py
