+++ src/maccabipediabot/common/prettify_games_pages.py
