+++ src/maccabipediabot/football/gamesbot.py
@@ -3,7 +3,7 @@ import re
 from datetime import timedelta
 from typing import AnyStr, List
 
-from maccabipediabot.pywikibot_boilerplate import run_boilerplate
+from maccabipediabot.common.pywikibot_boilerplate import run_boilerplate
 
 run_boilerplate()
 
@@ -18,9 +18,9 @@ site.login()
 from maccabistats import get_maccabi_stats_as_newest_wrapper
 from maccabistats.models.player_game_events import GameEventTypes
 from maccabistats.stats.maccabi_games_stats import MaccabiGamesStats
-from maccabipediabot.maccabistats_player_event import PlayerEvent
-from maccabipediabot.prettify_games_pages import prettify_game_page_main_template
-from maccabipediabot.sort_players_events import sort_player_events_in_games_page
+from maccabipediabot.common.maccabistats_player_event import PlayerEvent
+from maccabipediabot.common.prettify_games_pages import prettify_game_page_main_template
+from maccabipediabot.football.sort_players_events import sort_player_events_in_games_page
 
 logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)
