+++ src/maccabipediabot/maintenance/move_table_template.py
@@ -2,7 +2,7 @@ import logging
 
 from pywikibot import pagegenerators, Category
 
-from maccabipediabot.pywikibot_boilerplate import run_boilerplate
+from maccabipediabot.common.pywikibot_boilerplate import run_boilerplate
 
 logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.DEBUG)
